requirements.txt
Flask
Pillow
fpdf
PyPDF2
